+++ continual/convert.py
@@ -11,7 +11,7 @@ from .container import Sequential
 from .conv import Conv1d, Conv2d, Conv3d
 from .linear import Linear
 from .logging import getLogger
-from .module import CallMode, CoModule, call_mode
+from .module import CoModule, _callmode, call_mode
 from .pooling import (
     AdaptiveAvgPool2d,
     AdaptiveAvgPool3d,
@@ -46,6 +46,13 @@ def forward_stepping(module: nn.Module, dim: int = 2):
         dim (int, optional): The dimension to unsqueeze during `forward_step`. Defaults to 2.
     """
 
+    def _forward_step(func: Callable[[Tensor], Tensor]):
+        @wraps(func)
+        def call(x: Tensor, prev_state=None) -> Tensor:
+            return func(x.unsqueeze(dim)).squeeze(dim), prev_state
+
+        return call
+
     def forward_step(func: Callable[[Tensor], Tensor]):
         @wraps(func)
         def call(x: Tensor, update_state=True) -> Tensor:
@@ -72,25 +79,28 @@ def forward_stepping(module: nn.Module, dim: int = 2):
     orig_forward = module.forward
     module.forward_steps = forward_steps(module.forward)
     module.forward_step = forward_step(module.forward)
+    module._forward_step = _forward_step(module.forward)
     module.delay = 0
     module.receptive_field = 1
-    module.stride = getattr(module, "stride", 1)
-    module.padding = getattr(module, "padding", 0)
+    module.stride = tuple(getattr(module, "stride", [1]))
+    module.padding = tuple(getattr(module, "padding", [0]))
     module.build_from = build_from
     module.get_state = dummy
     module.set_state = dummy
     module.clean_state = dummy
+    module._state_shape = 0
+    module._dynamic_state_inds = []
 
     # Call mode
-    module.call_mode = CallMode.FORWARD
+    module.call_mode = _callmode("forward")
 
     def forward_with_callmode(*args, **kwargs):
         _call_mode = (
             call_mode.cur
-            if call_mode.prev
-            else getattr(module, "call_mode", CallMode.FORWARD)
+            if call_mode.prev is not None
+            else getattr(module, "call_mode", _callmode("forward"))
         )
-        if _call_mode == CallMode.FORWARD:
+        if _call_mode == _callmode("forward"):
             return orig_forward(*args, *kwargs)
         return CoModule._call_impl(module, *args, **kwargs)