+++ continual/delay.py
@@ -3,10 +3,10 @@ from typing import Tuple, Union
 import torch
 from torch import Tensor
 
-from .module import CoModule, PaddingMode, TensorPlaceholder
+from .module import CoModule, PaddingMode
 from .utils import temporary_parameter
 
-State = Tuple[Tensor, int]
+State = Tuple[Tensor, Tensor]
 
 __all__ = ["Delay"]
 
@@ -17,6 +17,14 @@ class Delay(CoModule, torch.nn.Module):
     This module only introduces a delay in the continual modes, i.e. on `forward_step` and `forward_steps`.
     This corresponds to the equvalent computations when delay is used to align continual computations.
     """
+
+    @property
+    def _state_shape(self):
+        return 2 if self.delay > 0 else 0
+
+    @property
+    def _dynamic_state_inds(self):
+        return [True, False] if self.delay > 0 else []
 
     def __init__(
         self,
@@ -43,7 +51,8 @@ class Delay(CoModule, torch.nn.Module):
         ]
 
         super(Delay, self).__init__()
-        # state is initialised in self.forward
+        self.register_buffer("state_buffer", torch.tensor([]), persistent=False)
+        self.register_buffer("state_index", torch.tensor(0), persistent=False)
 
     def init_state(
         self,
@@ -51,30 +60,25 @@ class Delay(CoModule, torch.nn.Module):
     ) -> State:
         padding = self.make_padding(first_output)
         state_buffer = torch.stack([padding for _ in range(self.delay)], dim=0)
-        state_index = -self.delay
-        if not hasattr(self, "state_buffer"):
-            self.register_buffer("state_buffer", state_buffer, persistent=False)
+        state_index = torch.tensor(-self.delay)
         return state_buffer, state_index
 
     def clean_state(self):
-        if hasattr(self, "state_buffer"):
-            del self.state_buffer
-        if hasattr(self, "state_index"):
-            del self.state_index
+        self.state_buffer = torch.tensor([])
+        self.state_index = torch.tensor(0)
 
     def get_state(self):
-        if (
-            hasattr(self, "state_buffer")
-            and self.state_buffer is not None
-            and hasattr(self, "state_index")
-            and self.state_buffer is not None
-        ):
+        if len(self.state_buffer) > 0:
             return (self.state_buffer, self.state_index)
+        return None
 
     def set_state(self, state: State):
         self.state_buffer, self.state_index = state
 
     def _forward_step(self, input: Tensor, prev_state: State) -> Tuple[Tensor, State]:
+        if self._delay == 0:
+            return input, prev_state
+
         if prev_state is None:
             buffer, index = self.init_state(input)
         else:
@@ -84,7 +88,7 @@ class Delay(CoModule, torch.nn.Module):
         if index >= 0:
             output = buffer[index].clone()
         else:
-            output = TensorPlaceholder(buffer[0].shape)
+            output = None
 
         # Update state
         buffer[index % self.delay] = input
@@ -129,49 +133,9 @@ class Delay(CoModule, torch.nn.Module):
         return self._delay
 
     @property
-    def stride(self) -> int:
-        return 1
+    def stride(self) -> Tuple[int]:
+        return (1,)
 
     def extra_repr(self):
         shrink_str = ", auto_shrink=True" if self.auto_shrink else ""
         return f"{self.delay}" + shrink_str
-
-
-class DiscardFirstSteps(CoModule, torch.nn.Module):
-    """Discard when data is first passed to module"""
-
-    def __init__(
-        self,
-        num_steps: int,
-    ):
-        assert num_steps >= 0
-        self._num_steps = num_steps
-        super(DiscardFirstSteps, self).__init__()
-
-    def init_state(
-        self,
-        first_output: Tensor,
-    ) -> State:
-        state_index = -self._num_steps
-        return state_index
-
-    def clean_state(self):
-        if hasattr(self, "state_index"):
-            del self.state_index
-
-    def get_state(self):
-        if hasattr(self, "state_index") and self.state_buffer is not None:
-            return self.state_index
-
-    def set_state(self, state: State):
-        self.state_index = state
-
-    def _forward_step(self, input: Tensor, prev_state: State) -> Tuple[Tensor, State]:
-        index = prev_state or self.init_state(input)
-        if index >= 0:
-            return input, index
-        else:
-            return TensorPlaceholder(input.shape), index + 1
-
-    def forward(self, input: Tensor) -> Tensor:
-        return input[:, :, -self._num_steps :].unsqueeze(2)