+++ continual/multihead_attention/retroactive_mha.py
@@ -1,12 +1,12 @@
 import math
 from functools import partial
 from logging import getLogger
-from typing import Optional, Tuple, Union
+from typing import Optional, Tuple
 
 import torch
 from torch import Tensor
 
-from continual.module import CallMode, TensorPlaceholder
+from continual.module import _callmode
 
 from .mha_base import MultiheadAttentionBase, scaled_dot_prod_attn_flops
 
@@ -18,6 +18,7 @@ State = Tuple[
     Tensor,  # Q_mem, (B, Nt-1, E)
     Tensor,  # K_T_mem, (B, E, Ns)
     Tensor,  # V_mem, (B, Ns, E)
+    Tensor,  # state_index
 ]
 
 
@@ -172,6 +173,10 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
     number of features.
     """
 
+    _state_shape = 6
+    #                     d_mem, AV_mem, Q_mem, K_T_mem, V_mem ,state_index
+    _dynamic_state_inds = [True, True, True, True, True, False]
+
     def __init__(
         self,
         embed_dim,
@@ -213,6 +218,12 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
             forward_returns_attn_mask,
             embed_dim_second,
         )
+        self.register_buffer("d_mem", torch.tensor([]), persistent=False)
+        self.register_buffer("AV_mem", torch.tensor([]), persistent=False)
+        self.register_buffer("Q_mem", torch.tensor([]), persistent=False)
+        self.register_buffer("K_T_mem", torch.tensor([]), persistent=False)
+        self.register_buffer("V_mem", torch.tensor([]), persistent=False)
+        self.register_buffer("stride_index", torch.tensor(0), persistent=False)
 
     def get_state(self) -> Optional[State]:
         """Get model state
@@ -220,14 +231,7 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
         Returns:
             Optional[State]: A State tuple if the model has been initialised and otherwise None.
         """
-        if (
-            getattr(self, "d_mem", None) is not None
-            and getattr(self, "AV_mem", None) is not None
-            and getattr(self, "Q_mem", None) is not None
-            and getattr(self, "K_T_mem", None) is not None
-            and getattr(self, "V_mem", None) is not None
-            and getattr(self, "stride_index", None) is not None
-        ):
+        if len(self.d_mem) > 0:
             return (
                 self.d_mem,
                 self.AV_mem,
@@ -254,18 +258,54 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
 
     def clean_state(self):
         """Clean model state"""
-        if hasattr(self, "d_mem"):
-            del self.d_mem
-        if hasattr(self, "AV_mem"):
-            del self.AV_mem
-        if hasattr(self, "Q_mem"):
-            del self.Q_mem
-        if hasattr(self, "K_T_mem"):
-            del self.K_T_mem
-        if hasattr(self, "V_mem"):
-            del self.V_mem
-        if hasattr(self, "stride_index"):
-            del self.stride_index
+        self.d_mem = torch.tensor([])
+        self.AV_mem = torch.tensor([])
+        self.Q_mem = torch.tensor([])
+        self.K_T_mem = torch.tensor([])
+        self.V_mem = torch.tensor([])
+        self.stride_index = torch.tensor(0)
+
+    def _forward_step(
+        self,
+        query: Tensor,
+        key: Tensor = None,
+        value: Tensor = None,
+        prev_state: Optional[State] = None,
+    ) -> Tuple[Optional[Tensor], State]:
+        """
+        Args:
+            query, key, value: step_inputs for mapping a query and a set of key-value pairs to an output.
+                See "Attention Is All You Need" for more details.
+
+        Shapes for inputs:
+            - query: :math:`(N, E)` where L is the target sequence length, N is the batch size, E is
+              the embedding dimension.
+            - key: :math:`(N, E)`, where S is the source sequence length, N is the batch size, E is
+              the embedding dimension.
+            - value: :math:`(N, E)` where S is the source sequence length, N is the batch size, E is
+              the embedding dimension.
+
+        Shapes for outputs:
+            - attn_output: :math:`(L, N, E)` where L is the target sequence length, N is the batch size,
+              E is the embedding dimension. :math:`(N, L, E)` if ``batch_first`` is ``True``.
+              :math:`(N, E, L)` if ``batch_first`` and ``embed_dim_second ``True``.
+        """
+        if key is None:
+            key = query
+        if value is None:
+            value = query
+
+        o, next_state = MultiheadAttentionBase._forward_step(
+            self, query, key, value, prev_state
+        )
+
+        if o is not None:
+            if self.batch_first:
+                o = o.transpose(1, 0)
+            if self.embed_dim_second:
+                o = o.transpose(1, 2)
+
+        return o, next_state
 
     def forward_step(
         self,
@@ -275,7 +315,7 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
         update_state=True,
         *args,
         **kwargs,
-    ) -> Union[Tensor, TensorPlaceholder]:
+    ) -> Optional[Tensor]:
         """
         Args:
             query, key, value: step_inputs for mapping a query and a set of key-value pairs to an output.
@@ -311,7 +351,7 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
         update_state=True,
         *args,
         **kwargs,
-    ) -> Union[Tensor, TensorPlaceholder]:
+    ) -> Optional[Tensor]:
         """Forward computation for multiple steps with state initialisation
 
         Args:
@@ -337,8 +377,8 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
 
         # Linear projection
         steps_taken = {
-            CallMode.FORWARD: self.sequence_len,
-            CallMode.FORWARD_STEP: 1,
+            _callmode("forward"): self.sequence_len,
+            _callmode("forward_step"): 1,
         }[self.call_mode]
 
         f += (
@@ -359,8 +399,8 @@ class RetroactiveMultiheadAttention(MultiheadAttentionBase):
 
         # Multi-head Scaled Dot-Product Attention
         f += self.num_heads * {
-            CallMode.FORWARD: scaled_dot_prod_attn_flops,
-            CallMode.FORWARD_STEP: retractive_scaled_dot_prod_attn_step_flops,
+            _callmode("forward"): scaled_dot_prod_attn_flops,
+            _callmode("forward_step"): retractive_scaled_dot_prod_attn_step_flops,
         }[self.call_mode](
             self.sequence_len,
             self.embed_dim // self.num_heads,