+++ tests/continual/multihead_attention/test_retroactive_mha.py
@@ -4,7 +4,6 @@ import torch
 from ptflops import get_model_complexity_info
 from torch.nn.modules.activation import MultiheadAttention
 
-from continual.module import TensorPlaceholder
 from continual.multihead_attention.retroactive_mha import (
     RetroactiveMultiheadAttention,
     _scaled_dot_product_attention_default_state,
@@ -59,7 +58,7 @@ def test_multi_head_attention():
     # Initialise, then use forward_step
     comha.clean_state()
     attn_output_dummy = comha.forward_steps(query[:-1], key[:-1], value[:-1])
-    assert isinstance(attn_output_dummy, TensorPlaceholder)
+    assert attn_output_dummy is None
     attn_output4 = comha.forward_step(query[-1], key[-1], value[-1])
     assert torch.allclose(attn_output, attn_output4, atol=1e-6)